Use file_contexts.bin first then use file_contexts if it not exist
Also remove useless file_contexts write lines of multirom.

+++ multirom/multirom.cpp
@@ -316,19 +316,35 @@ bool MultiROM::restorecon(std::string name)
 {
 	bool res = false;
 	bool replaced_contexts = false;
+	bool file_contexts_old = false;
 
 	std::string file_contexts = getRomsPath() + name;
 	std::string seapp_contexts = file_contexts + "/boot/seapp_contexts";
-	file_contexts += "/boot/file_contexts";
+	if (TWFunc::Path_Exists(file_contexts + "/boot/file_contexts.bin")) {
+		file_contexts += "/boot/file_contexts.bin";
 
-	if(access(file_contexts.c_str(), R_OK) >= 0)
-	{
-		gui_print("Using ROM's file_contexts\n");
-		rename("/file_contexts", "/file_contexts.orig");
-		rename("/seapp_contexts", "/seapp_contexts.orig");
-		system_args("cp -a \"%s\" /file_contexts", file_contexts.c_str());
-		system_args("cp -a \"%s\" /seapp_contexts", seapp_contexts.c_str());
-		replaced_contexts = true;
+		if(access(file_contexts.c_str(), R_OK) >= 0)
+		{
+			gui_print("Using ROM's file_contexts.bin\n");
+			rename("/file_contexts.bin", "/file_contexts.bin.orig");
+			rename("/seapp_contexts", "/seapp_contexts.orig");
+			system_args("cp -a \"%s\" /file_contexts.bin", file_contexts.c_str());
+			system_args("cp -a \"%s\" /seapp_contexts", seapp_contexts.c_str());
+			replaced_contexts = true;
+		}
+	} else if (TWFunc::Path_Exists(file_contexts + "/boot/file_contexts")) {
+		file_contexts += "/boot/file_contexts";
+
+		if(access(file_contexts.c_str(), R_OK) >= 0)
+		{
+			gui_print("Using ROM's file_contexts\n");
+			rename("/file_contexts", "/file_contexts.orig");
+			rename("/seapp_contexts", "/seapp_contexts.orig");
+			system_args("cp -a \"%s\" /file_contexts", file_contexts.c_str());
+			system_args("cp -a \"%s\" /seapp_contexts", seapp_contexts.c_str());
+			replaced_contexts = true;
+			file_contexts_old = true;
+		}
 	}
 
 	if(!changeMounts(name))
@@ -350,7 +366,11 @@ bool MultiROM::restorecon(std::string name)
 	res = true;
 exit:
 	if(replaced_contexts) {
-		rename("/file_contexts.orig", "/file_contexts");
+		if (file_contexts_old) {
+			rename("/file_contexts.orig", "/file_contexts");
+		} else {
+			rename("/file_contexts.bin.orig", "/file_contexts.bin");
+		}
 		rename("/seapp_contexts.orig", "/seapp_contexts");
 	}
 	return res;
@@ -1653,15 +1673,27 @@ bool MultiROM::extractBootForROM(std::string base)
 	}
 
 	// copy needed files
-	static const char *cp_f[] = {
-		"*.rc", "default.prop", "init", "main_init", "fstab.*",
-		// Since Android 4.3 - for SELinux
-		"file_contexts", "property_contexts", "seapp_contexts", "sepolicy",
-		NULL
-	};
+	if (TWFunc::Path_Exists("/tmp/boot/file_contexts.bin")) {
+		static const char *cp_f[] = {
+			"*.rc", "default.prop", "init", "main_init", "fstab.*",
+			// Since Android 4.3 - for SELinux
+			"file_contexts.bin", "property_contexts", "seapp_contexts", "sepolicy",
+			NULL
+		};
 
-	for(int i = 0; cp_f[i]; ++i)
-		system_args("cp -a /tmp/boot/%s \"%s/boot/\"", cp_f[i], base.c_str());
+		for(int i = 0; cp_f[i]; ++i)
+			system_args("cp -a /tmp/boot/%s \"%s/boot/\"", cp_f[i], base.c_str());
+	} else {
+		static const char *cp_f[] = {
+			"*.rc", "default.prop", "init", "main_init", "fstab.*",
+			// Since Android 4.3 - for SELinux
+			"file_contexts", "property_contexts", "seapp_contexts", "sepolicy",
+			NULL
+		};
+
+		for(int i = 0; cp_f[i]; ++i)
+			system_args("cp -a /tmp/boot/%s \"%s/boot/\"", cp_f[i], base.c_str());
+	}
 
 	// check if main_init exists
 	sprintf(path, "%s/boot/main_init", base.c_str());
